refactor: separate event binding into eventView and extract playLottoGameWeb controller
